backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.19.1
redis[hiredis]==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
aioredis==2.0.1
